refactor: code locations
