+++ src/teamster/kippcamden/deanslist/config/multi-partition-monthly-assets.yaml
@@ -1,5 +1,5 @@
 endpoints:
-  - asset_name: behavior
+  - endpoint: behavior
     api_version: v1
     params:
       IncludeDeleted: Y
@@ -9,11 +9,11 @@ endpoints:
           resources:
             requests:
               memory: 3.0Gi
-  - asset_name: homework
+  - endpoint: homework
     api_version: v1
     params:
       IncludeDeleted: Y
-  - asset_name: incidents
+  - endpoint: incidents
     api_version: v1
     params:
       IncludeDeleted: Y
