+++ src/teamster/kippmiami/deanslist/config/multi-partition-fiscal-assets.yaml
@@ -1,8 +1,8 @@
 endpoints:
-  - asset_name: comm-log
+  - endpoint: comm-log
     api_version: v1
     params:
       IncludeDeleted: Y
       IncludePrevEnrollments: Y
-  - asset_name: followups
+  - endpoint: followups
     api_version: v1
