+++ src/teamster/kippmiami/deanslist/config/multi-partition-monthly-assets.yaml
@@ -1,13 +1,13 @@
 endpoints:
-  - asset_name: behavior
+  - endpoint: behavior
     api_version: v1
     params:
       IncludeDeleted: Y
-  - asset_name: homework
+  - endpoint: homework
     api_version: v1
     params:
       IncludeDeleted: Y
-  - asset_name: incidents
+  - endpoint: incidents
     api_version: v1
     params:
       IncludeDeleted: Y
