+++ src/teamster/kippmiami/deanslist/config/static-partition-assets.yaml
@@ -1,19 +1,19 @@
 endpoints:
-  - asset_name: users
+  - endpoint: users
     api_version: v1
     params:
       IncludeInactive: Y
-  - asset_name: lists
+  - endpoint: lists
     api_version: v1
-  - asset_name: terms
+  - endpoint: terms
     api_version: v1
-  - asset_name: rosters
+  - endpoint: rosters
     api_version: v1
     params:
       show_inactive: Y
-  - asset_name: roster-assignments
+  - endpoint: roster-assignments
     api_version: beta
-  - asset_name: students
+  - endpoint: students
     api_version: v1
     params:
       IncludeCustomFields: Y
